requirements.txt
asttokens==2.0.8
backcall==0.2.0
decorator==5.1.1
executing==1.0.0
ipdb==0.13.9
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
jinja2-markdown==0.0.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
PyYAML==6.0
six==1.16.0
stack-data==0.5.0
toml==0.10.2
traitlets==5.4.0
wcwidth==0.2.5
